requirements.txt
Flask==3.0.2
gunicorn==21.2.0
google-generativeai==0.5.4
Pillow==10.3.0
Werkzeug==3.0.1
